+++ lib/store/helpers/elements/index.ts
@@ -1,4 +1,7 @@
 import { getFormElementById } from "./getFormElementById";
 import { getFormElementWithIndexById } from "./getFormElementWithIndexById";
+import { propertyPath } from "./propertyPath";
+import { getPathString } from "./getPathString";
+import { getChoice } from "./getChoice";
 
-export { getFormElementById, getFormElementWithIndexById };
+export { getFormElementById, getFormElementWithIndexById, propertyPath, getPathString, getChoice };
